requirements.txt
asgiref==3.3.4
Django==3.2.3
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
Markdown==3.3.4
Pillow==8.2.0
PyJWT==2.1.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
